fix(api): add error handling for missing access token in EbayApiClient
This commit introduces a check in the EbayApiClient to throw an error if the access token is missing, providing a clear message to the user about the required credentials. Additionally, it refactors imports across various files to streamline the codebase, including the reorganization of tool definitions and the introduction of a new types module for better type management.

Key changes include:
- Added error handling for missing access token in `client.ts`.
- Refactored imports in `shared.ts`, `inventory.ts`, and other definition files to use the new `types.ts`.
- Consolidated tool definitions into a new `connector.ts` file for better organization.

These changes aim to enhance the robustness and maintainability of the API client.
