+++ src/tools/definitions/developer.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import type { OutputArgs, ToolDefinition } from '../tool-definitions.js';
+import type { OutputArgs, ToolDefinition } from './types.js';
 
 /** Developer API tools for eBay application and keyset management. */
 export const developerTools: ToolDefinition[] = [
