+++ src/tools/index.ts
@@ -1,5 +1,5 @@
 /** Public tool definition type exported by the tools module. */
-export type { ToolDefinition } from '@/tools/tool-definitions.js';
+export type { ToolDefinition } from '@/tools/definitions/types.js';
 
 /** Public tool contract helpers exported by the tools module. */
 export { getToolContracts, validateToolContracts, type ToolContract } from '@/tools/contracts.js';
